backend/pyproject.toml
[project]
name = "estudaqui"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pdfplumber (>=0.11.6,<0.12.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pymupdf (>=1.26.0,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[tool.poetry]
packages = [{include = "estudaqui", from = "src"}]


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
